Rename sitters page to bookings page
