+++ client/src/App.tsx
@@ -12,7 +12,7 @@ import { AuthProvider } from './context/useAuthContext';
 import { SocketProvider } from './context/useSocketContext';
 import { SnackBarProvider } from './context/useSnackbarContext';
 import { Navbar } from './components/Navbar/Navbar';
-import Sitters from './pages/Sitters/Sitters';
+import Bookings from './pages/Bookings/Bookings';
 import Settings from './pages/Settings/Settings';
 import NotFound from './pages/NotFound/NotFound';
 
@@ -30,7 +30,7 @@ function App(): JSX.Element {
                 <Route exact path="/login" component={Login} />
                 <Route exact path="/signup" component={Signup} />
                 <Route exact path="/dashboard" component={Dashboard} />
-                <Route path="/sitters" component={Sitters} />
+                <Route path="/my-jobs" component={Bookings} />
                 <Route path="/profile/settings" component={Settings} />
                 <Route path="*">
                   <NotFound />
